Move remaining math tests out of test_geom
